crawler/requirements.txt
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.3.0
pandas==2.2.3
firebase-admin==6.5.0
python-dotenv==1.0.1
schedule==1.2.2
